pyproject.toml
[tool.poetry]
name = "jushma-test"
version = "0.1.0"
description = ""
authors = ["aSel1x <https://t.me/aSel1x>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
litestar = {version = "^2", extras = ["standard"]}
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
litestar-granian = "^0"
litestar-asyncpg = "^0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"